requirements.txt
fastapi~=0.115.4
uvicorn==0.24.0
python-multipart==0.0.7
pydantic~=2.9.2
python-dotenv==1.0.0
motor==3.3.1
assemblyai==0.17.0
requests==2.32.0
python-jose==3.3.0
MyApplication~=0.1.0
jose~=1.0.0
passlib~=1.7.4
pillow~=10.3.0
aiohttp~=3.11.2
pydub~=0.25.1
dnspython~=2.7.0
idna~=3.10
httpcore~=1.0.7
httpx~=0.27.2
anyio~=4.6.2.post1
pip~=24.3.1
jupyter~=1.1.1
attrs~=24.2.0
distro~=1.9.0
wheel~=0.44.0
filelock~=3.13.1
redis~=5.2.0
pymongo~=4.10.1
tornado~=6.4.2
certifi~=2024.8.30